grpc-demo/api_gateway/connection_polling/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
grpcio>=1.50.0
protobuf>=4.0.0
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
